When running (unconfined) code with rune, give it a <this> loader by default
If you run "rune file.e", <this> will allow importing other E source code from
the same directory as "file.e". Files imported this way will also get a
reference to <this>.

If you use --src or "rune -", this will load files from the current directory.
This is useful for interactive use.

If you run "rune file.emaker" (or use --confine), the <this> loader will not
be available.