Split Scope into Scope and EEnv
Currently, they're identical and calling EEnv.asScope() gets a Scope from an
EEnv. EExpr.eval and compile now take an EEnv, not a Scope.

The idea is to separate out the parts of Scope that are part of the E language,
and the parts that are internal to the interpreter.