Relax ordering of Ref.whenResolved to improve efficiency
Before, it didn't execute the when block until all messages to the target
had been processed, which required an extra network round-trip in the case
where the target was already resolved, and extra messages and processing
in the case where it was a promise.

Now, whenResolved is handled specially, so that the callback is invoked as
soon as the target is known to be resolved.